pinning rows, and context menu improvements
